Bluetooth: controller: Send disconnect complete when CIS is terminated
Previously the host would handle disconnecting associated CISes and
invoking callbacks when a host ACL conn was disconnected.

This responsibility has now been moved to the controller, which needs to
send a NODE_RX_TYPE_TERMINATE with the CIS handle for proper
disconnection in the host. This is in accordance with the spec.

As disconnect reason, the CIS uses the ACL reason, which is passed to
the host along with the handle.

(cherry picked from commit cd470da214c90328a801dd67f115eae4ffa5c988)

Signed-off-by: Morten Priess <mtpr@oticon.com>
GitOrigin-RevId: cd470da214c90328a801dd67f115eae4ffa5c988
Change-Id: Idf0a69ccc6b375594d253c709d510ab4fdc27ed2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3250772
Tested-by: CopyBot <cros.copybot.ota@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>