drivers: can: flexcan: fix timing parameter limits
Fix the limits for the timing parameter calculations.

The lower limit for the phase_seg2 value is wrongly specified as 1 to 7,
but 1U is substracted before writing it to the CTRL1:PSEG2 register
field. This results in register field values between 0 and 6, but 0 is
an invalid value for the PSEG2 register field.

The upper limits for several of the timing parameters are wrong as well,
but this does not result in invalid register field values being
calculated. It can, however, result in not being able to meet CAN timing
requirements.

The confusion in specifying the limits likely stems from the timing
calculations and timing limits using the "physical" values, whereas the
registers fields all use the "physical" value minus 1. When the
datasheet says "The valid programmable values are 1-7", the
corresponding limits should be set to 2 to 8 to take the "minus 1" into
account.

(cherry picked from commit 267a83e94b15adb6fe5d95b1eb46455c17af9d18)

Fixes: #39541
Signed-off-by: Henrik Brix Andersen <hebad@vestas.com>
GitOrigin-RevId: 267a83e94b15adb6fe5d95b1eb46455c17af9d18
Change-Id: I801b77454cc914235ba47b5014e602bfc105ee03
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3234021
Tested-by: CopyBot <cros.copybot.ota@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>