drivers: can: flexcan: fix timing parameter limits
Fix the limits for the timing parameter calculations.

The lower limit for the phase_seg2 value is wrongly specified as 1 to 7,
but 1U is substracted before writing it to the CTRL1:PSEG2 register
field. This results in register field values between 0 and 6, but 0 is
an invalid value for the PSEG2 register field.

The upper limits for several of the timing parameters are wrong as well,
but this does not result in invalid register field values being
calculated. It can, however, result in not being able to meet CAN timing
requirements.

The confusion in specifying the limits likely stems from the timing
calculations and timing limits using the "physical" values, whereas the
registers fields all use the "physical" value minus 1. When the
datasheet says "The valid programmable values are 1-7", the
corresponding limits should be set to 2 to 8 to take the "minus 1" into
account.

(cherry picked from commit 5166ff9fb1a51b7b12f0ff8e49b2cf412c1f062a)

Fixes: #39541
Signed-off-by: Henrik Brix Andersen <hebad@vestas.com>
GitOrigin-RevId: 5166ff9fb1a51b7b12f0ff8e49b2cf412c1f062a
Change-Id: Ibce9bde9c421a4446711c2753111a7af666ab55c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3236838
Tested-by: CopyBot <cros.copybot.ota@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>