soc/intel_adsp: Unify "RAM" region
There was a divergent set of definitions for a "RAM" region for the
benefit of memory layout in the platform headers.  In fact there was
only one platform dependence (cAVS 1.5 has 32k instead of 64k
reserved).  Put that into kconfig in a single place, and add a warning
that this is a trap region with hidden dependencies in both Zephyr and
SOF.  Good enough until we clean this up and make everything visible
to the linker.

(cherry picked from commit 33bd649d4e872a02d8b3a2ef6e6d1fe269dfc3aa)

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: 33bd649d4e872a02d8b3a2ef6e6d1fe269dfc3aa
Change-Id: I695c2c577685c3d5a36782fc2d237924e69383e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3300233
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>