soc/intel_adsp: Unify "firmware status register" window size
There is a diverged definition for SRAM_REG_FW_END, which exists to
prevent the Zephyr window initialization from changing values that
were already set by the ROM or bootloader (though this is incomplete,
as we're not ensuring the memory is actually the same space except by
convention; we also don't have any Zephyr-side visibility as to the
content of this struct).

That was silly; the only thing worse than one magic number is four
magic numbers in different files.  Write a formula that works for all
the platforms and put it in the C file where it's used.

(cherry picked from commit 003c878245a2e7dcfa59531087cbcddb000e5f85)

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: 003c878245a2e7dcfa59531087cbcddb000e5f85
Change-Id: I6eaf68033f2626e0752de7e4854990ca69b25d76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3300234
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>