Bluetooth: ISO: Store SDU and PHY for peripheral
When an CIS is connected, the peripheral did not have
any information about the QoS settings. This commit
adds information about the PHY and SDU. For some reason
the peripheral won't ever have information about the RTN.

The remaining values in the event (interval, delay and
latency) are still not exposed, nor is framing or
packing, the latter of which are not part of the event.

(cherry picked from commit 69f1236657e01a496e85a85223ac763d63db4b77)

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 69f1236657e01a496e85a85223ac763d63db4b77
Change-Id: Ia3370f074dfee5601bea4e5eb1d05eb3353da9d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3267256
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>