drivers: flash : introduce a new driver for the OCTOSPI
This is the stm32 octospi driver based on the exisitng quadspi
 for stm32 devices and source code from the STM32Cube.
This drivers initialized the peripheral and the NOR memory
in SPI or OctoSPI mode with STR or DTR data Transfer rates.
The NOR-flash can provide the SDFP table directly (if supported)
or through the DeviceTree.
 Limitation: no DMA transfer.

Signed-off-by: Francois Ramu <francois.ramu@st.com>
GitOrigin-RevId: 16754378fb8c16c0242f129f54deff5fb918e574
Change-Id: Iffdb720b812df691dd3f24f302d69e31f8223e13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3647568
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>