net: openthread: Change OT net log to display current role in text
If CONFIG_NET_LOG=y is set, OpenThread will output the current OT
role whenever the state changes.

To simplify understanding of the log output, this change replaces
the numerical role ID with the text name of the role. This also
required a change to a documentation file to replace an instance
of a numerical ID.

test scripts that monitor the OpenThread state changes and look
for the numerical ID. This does not seem to be the case for the
Zephyr tests, however.

(cherry picked from commit b8dff131ddd5ede81f6409957261e922354bf524)

NOTE: This is potentially a breaking change should anyone be using
Signed-off-by: Chris Pearson <ctpearson@gmail.com>
GitOrigin-RevId: b8dff131ddd5ede81f6409957261e922354bf524
Change-Id: I93524c61bd4c8f5501d43f9c4d5df90bc400e315
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3257723
Tested-by: CopyBot <cros.copybot.ota@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>