drivers: spi_nrfx_spis: Fix handling of buffer lengths
Do not limit the length of the prepared transfer to the shorter of
provided TX and RX buffers if both are used. The SPIS peripheral
cannot handle scattered buffers anyway, so there is no point in
getting the common part of TX and RX buffers for a partial transfer,
like it is done for SPI and SPIM peripherals; everything what is
possible needs to be transferred in one shot. For the same reason,
there is no point in calling spi_context_buffers_setup() and using
the related part of the spi_context structure, hence the call is
removed and buffer pointers are used directly.
Also return an error if a requested transfer length exceeds the SPIS
peripheral hardware limit, instead of silently limiting the transfer
like it was done so far.

(cherry picked from commit eafafa2816503673fc56aee9b9925fd59bf2cb8e)

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: eafafa2816503673fc56aee9b9925fd59bf2cb8e
Change-Id: I9ae6cd51c852dadb84265daa41c43aa95ad5584d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3262225
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Tristan Honscheid <honscheid@google.com>