arch/x86_64: Add hook for CONFIG_SCHED_THREAD_USAGE accounting in ISRs
Call into z_thread_usage_stop() before ISR entry to avoid including
interrupt handling totals in thread usage stats.

This has to go into the assembly immediately before the callback-based
dispatch.  Note that the dispatch code was putting the vector number
in RCX, which was unfortunate as that's a caller-saved register.
Would be nice to clean this up in the future so it lives in a
preserved register but it's mildly complicated to make work with the
way we do the stack layout right now.

(cherry picked from commit 1238410914597aef1e0661e0b2e240a448f2dd60)

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: 1238410914597aef1e0661e0b2e240a448f2dd60
Change-Id: If35ae07f6c17e481247bad068a629d38e0d2bf07
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3270265
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>