arch: arm: core: aarch32: Fix the syscall design for Cortex-R
When calling a syscall, the SVC routine will now elevate the thread to
privileged mode and exit the SVC setting the return address to the
syscall handler.  When the thread is swapped back in, it will be running
z_do_arm_syscall in system mode.  That function will run the syscall
then automatically return the thread to usr mode.

This allows running the syscall in sys mode on a thread so that we can
use syscalls that sleep without doing unnatural things.  The previous
implementation would enable interrupts while still in the SVC call and
do weird things with the nesting count.  An interrupt could happen
during this time when the syscall was still in the exception state, but
the nested count had been decremented too soon.  Correctness of the
nested count is important for future floating point unit work.

The Cortex-R behavior now matches that of Cortex-M.

(cherry picked from commit 2a357e5dfd749b9d7aa5bc5ff5ed8a3ef5e5a498)

Signed-off-by: Bradley Bolen <bbolen@lexmark.com>
GitOrigin-RevId: 2a357e5dfd749b9d7aa5bc5ff5ed8a3ef5e5a498
Change-Id: If7c18c245c98aab4ba346e63ade245b2bfb257f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3450376
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>