doc: restore environment variables page
This was moved into the application development guide in commit
bf9ceac ("doc: move env_vars into application development").

That commit claimed the following rationale:

  No reason why we are duplicating this information in a standalone
  guide when we already talk about env variables in the application
  developer guide.

A github comment later says the duplication was in "Important build
system variables":

This is incorrect and a misreading of the application development
guide.

The application development guide is talking about **cmake**
variables. These are not the same thing at all as **environment**
variables; treating them as if they are the same is simply wrong and
misleading, especially the "option 4" part that is really about
CMake variables, not environment variables.

In fact, the separate environment variables page was created for a
reason, which was to have a single place to explain to people unused
to managing their environment how things work, without cluttering
things up for people who already understand this.

There is nothing wrong with small, self-contained guides and in fact I
would argue they are a good thing.

Fix this confusing situation by restoring the original, separate page
and extricating the portions of it that are not related to environment
variables.

(cherry picked from commit 71f17bcbc245a5648a8d07cf1ba65ef6e9b5cddf)

https://github.com/zephyrproject-rtos/zephyr/pull/34364#discussion_r616646517
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
GitOrigin-RevId: 71f17bcbc245a5648a8d07cf1ba65ef6e9b5cddf
Change-Id: I0bfb753ccd73f91a1b504f559daf76d88f099d94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3341696
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>