pm: Make pm_power_state_force multicore aware
Change pm_power_state_force to receive which cpu the state should be
forced. Also, it changed the API behavior to force the given state only
when the idle thread for that core is executed.

In a multicore environment force arbitrarily a core to suspend is not
safe because the kernel cannot infer what that cpu is running and how it
impacts the overall system, for example, if it is holding a lock that is
required by a thread that is running in another cpu.

(cherry picked from commit 4998c52ba8272061329648f95576bab59d68ae4f)

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>
GitOrigin-RevId: 4998c52ba8272061329648f95576bab59d68ae4f
Change-Id: I49a35963bc27b4553d82b6d525a273cce3a4ba94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3292526
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>