BACKPORT: pm: Make pm_power_state_force multicore aware
Change pm_power_state_force to receive which cpu the state should be
forced. Also, it changed the API behavior to force the given state only
when the idle thread for that core is executed.

In a multicore environment force arbitrarily a core to suspend is not
safe because the kernel cannot infer what that cpu is running and how it
impacts the overall system, for example, if it is holding a lock that is
required by a thread that is running in another cpu.

BUG=b:200317696
TEST=none

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>
(cherry picked from commit 4998c52ba8272061329648f95576bab59d68ae4f)
Change-Id: I5ae210d329242fe96c4b9445903d6366c269c434
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3302561
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Denis Brockus <dbrockus@chromium.org>
Tested-by: Michał Barnaś <mb@semihalf.com>
Auto-Submit: Michał Barnaś <mb@semihalf.com>
Commit-Queue: Denis Brockus <dbrockus@chromium.org>