Tracing: Fixed TraceRecorder inclusion problem
Moved conditional inclusion of tracing_tracerecorder.h
from the start of tracing.h to the bottom. This was the
initial placement of this include but it was moved as
part of PR #40126. If this file is included at the
start of tracing.h the TraceRecorder module hooks
will not map correctly and completely breaks
that module.

(cherry picked from commit e955b6540dde2ba41c67e6e94498bb2c6c76c082)

Signed-off-by: Torbjörn Leksell <torbjorn.leksell@percepio.com>
GitOrigin-RevId: e955b6540dde2ba41c67e6e94498bb2c6c76c082
Change-Id: I1402d7acec245aeea38111990d8f4c8b1c62e799
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3401202
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>