modules: trusted-firmware-m: Fix mcuboot imgtool not found on path
Fix issue with TFM signing of images not using the correct imgtool.
The wrapper command expects the mcuboot scripts folder to be the
current working directory when called in order to find its own
version of imgtool.
Since the command is using a different current working directory
this is not found and the system imgtool is used instead.
This causes the commands to be run with 2 different version of imgtool
if the system imgtool is found and does not have any issues.
The system imgtool could not be installed or have compatibility issues
as 1.7.2 version of imgtool is currently required by the wrapper script

(cherry picked from commit 13230747e3376bbaa4e521a7d181a9dfadca3af5)

Fixes: #40254
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
GitOrigin-RevId: 13230747e3376bbaa4e521a7d181a9dfadca3af5
Change-Id: I1f0f48dbcd30d39a18de3c4c7020b27e1499392e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3295662
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>