CHROMIUM: Temporarily revert atomic_t type refactor
Upstream Zephyr decided to convert the type of an atomic_t to alias long
instead of int, meaning 64-bit platforms can use 64-bit native atomics.

Due to the historical fact that the Chromium EC used to use uint32_t as
the atomic type, we have a significant amount of code in our tree which
assumes that uint32_t is compatible with atomic_t, which can no longer
be held true with the long type.

This temporarily reverts the refactor upstream buying us a little bit of
time to work on fixing up our tree to use atomic_t.  The plan is to
revert this commit after we've done the legwork in our tree to use long
universally as the atomic type.

The commits reverted are:

3e86c627: "kernel: atomics: update print specifiers for atomic_t"
174cb7f9: "kernel: atomics: support for 64-bit atomic operations"

BUG=b:207082842
TEST=zmake --zephyr-base <path_to_main_checkout> testall
     (build progresses much futher, but there's still other issues
      unrelated to resolve)

Change-Id: I615abee0e52eb742fcb232626bea0abb9629d98d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3293172
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>