cmake: extensions: add dt_alias()
This is a helper function for looking up the path to a devicetree
alias. It is analogous to dt_chosen(). It has to exist as a separate
function from dt_prop() for similar reasons that dt_chosen() does:
the edtlib APIs for interacting with /aliases are special-cased in the
same way they are for /chosen.

(cherry picked from commit a9f55781374e1832a8431bec2b4fdf7dcd5b02c8)

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
GitOrigin-RevId: a9f55781374e1832a8431bec2b4fdf7dcd5b02c8
Change-Id: I040d0fca405909788c2d714b8aee28c30560c438
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3318170
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>