soc/intel_adsp: Clean up IPM compiler guards
The CAVS_IDC_IPM driver happens to be used only on non-2.5 hardware,
but it's best to be clear in the conditional compilation when we're
talking about hardware-dependencies and when we mean software
configuration.  This was mixed up in a few spots.

Also fix a warning that creeps in on non-default drivers choices about
an undeclared ipm function.

(cherry picked from commit a7af0acfc28f1afc32c4b088f62260cd263a650a)

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: a7af0acfc28f1afc32c4b088f62260cd263a650a
Change-Id: I16547b10956df2dd915c70f27930be47f616f502
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3300246
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>