soc/intel_adsp: Simplify boot ROM protocol
There was an attempt in the old code to express this as a formal
protocol with a proper field definitions, etc...  But in fact no such
protocol really exists.  This scheme is only used in one place to send
one specific message to code fixed in ROM on legacy devices that only
knows how to recognize this specific value.  And 2.5 and later
hardware are moving away from it anyway.

Just express it directly, and explain in comments.

(cherry picked from commit c6de46c87e1f849243b742bccc61872dd071187f)

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: c6de46c87e1f849243b742bccc61872dd071187f
Change-Id: I219ea8badc6e4f2f8126089cc2ddd05985bca0fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3300247
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>