CHROMIUM: Add stub label for ite_uart1_wrapper
Since PR #39504, we need to enable ite_uart1_wrapper if we use uart1 in
device tree (similar for uart2, but we don't use it).

To allow the EC repo to enable the uart like this:

  &ite_uart1_wrapper {
          status = "okay";
  };

Add a small stub to the v2.7 branch only that'll die off with this
branch.

BUG=b:205884929
TEST=compile with uart enbaled on asurada for both v2.7 and main

Change-Id: I16278b459ca7c7d1e67b9311499afe19e15be6d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3293845
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Denis Brockus <dbrockus@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>