actions: twister/clang: cleanup test plan generation
Merged 3 files used to generate the test plan per PR based on the
changes. 2 python scripts and a shell script are now all merged into 1
python script that generates the input file for twister based on a list
of changed files by the PR.

This remove lots of old and obsolete code and simplifies things a bit,
no need anymore for an intermediate script to call twister, we call it
directly in the workflow and use the new test_plan script to generate
the test plan.

This also reenables the recently disabled tag based filtering which had
a bug, bug is resolved in this new implementation.

On push events, we now run twister without the --integration option to
catch any issues in the main branch that were not caught in PRs. PRs
continue to run with --integration enabled. This event (push) is now run
on 15 builders due to the increased size.

(cherry picked from commit 0b2df8f41c3a6679b241fc6e8a67cfcf48c99e43)

Signed-off-by: Anas Nashif <anas.nashif@intel.com>
GitOrigin-RevId: 0b2df8f41c3a6679b241fc6e8a67cfcf48c99e43
Change-Id: I4d52f52a668c792fe59773dee2cefbf67760d1d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3302402
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>