samples: tflite-micro: hello_world: Increase main stack size
This commit increases the main stack size to 2048 since the current
stack size is insufficient for the `qemu_x86` board (the default stack
size is 1024 for that board) and results in a stack overflow in the
main thread.

Note that this is required because the newlib full variant, which is
used by this sample, included in the Zephyr SDK 0.13.2 now has an
increased stack usage requirement due to enabling the C99 format
specifier support.

For more details, refer to the issue #40469 [1].

[1] https://github.com/zephyrproject-rtos/zephyr/pull/40469#issuecomment-977854095

(cherry picked from commit 6ee8182e0a434a756c6d45391cc5783406d372ac)

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: 6ee8182e0a434a756c6d45391cc5783406d372ac
Change-Id: I9d8766347e3a29bf42674d4bcb7a1ccfbf67183d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3302067
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>