cmake: introduce "snippet" concept
Introduce the concept of snippet's to application configuration. A
snippet is a piece of functionality that is not the default behaviour
of a board, but is a common configuration that users may wish to apply.

One example snippet concept is "usb" which may route the console to USB
instead of a serial port. Another common snippet could be "mcuboot",
which would define the flash partitions required for applications to
be loaded by a bootloader.

The name "snippet" was chosen mainly to avoid collisions with existing
terms. Alternatives considered include:
 * overlay (collision with devicetree overlay concept)
 * config (collision with Kconfig)
 * part (collision with --pristine build option, flash partitions)
 * layout (implies memory configuration)

The Oxford dictionary definition of snippet appears to match this
concept well, "a small piece or brief extract".

Snippets are a required devicetree `.overlay` file and an optional
`.conf` file that are appended to the existing CMake variables
`DTC_OVERLAY_FILE` and `CONF_FILE`. The initial search paths for these
files are `${APPLICATION_SOURCE_DIR}/${BOARD}/snippets` and
`${BOARD_DIR}/snippets`, with the application directory having
precedence to allow for overriding board defaults.

Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>