soc: arm: Add Aspeed ast10x0 series support
Aspeed ast10x0 series SoCs contain a ARM Cortex-M4F processor.  This
processor operates at 200MHz and executes on SRAM.

This patch adds support for ast1030 as the first SoC of this series
which is targeted but not limited at the bridge IC in a server system.

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Original-Change-Id: I668af1ff8a36a05da791c3329ae08f5ae712bdd4
GitOrigin-RevId: a24ab243221cb643c697a3b0ae25063fb6938246
Change-Id: I98484a3ca263f6300718ffa66fe5cb5b5b7cd251
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3551981
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>