soc/intel_adsp: Minimize bootloader linker script
The memory layout for the boot loader was needlessly complicated, with
separate fixed regions defined at fixed addresses, all in a file that
needs to be contiguous and DMA'd from the host in a single block.  The
end result was a lot of magic numbers and wasted space.

Clean things up so that it links in a single region expressed (for the
benefit of rimage, I think) as a single program header in the ELF
file.

This is in preparation for further changes to unify the bootloader
stage with the main Zephyr image in a single link.

(cherry picked from commit a42da8dcdb6eeccfefbcbb1782151d8d6cc6bb7b)

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: a42da8dcdb6eeccfefbcbb1782151d8d6cc6bb7b
Change-Id: I0ff736b0a5736812cf00bd126bc8dec3ce73a935
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3323284
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>